requirements.txt
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.44
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
openai==1.23.1
pinecone-client==3.2.2
python-dotenv==1.0.1
scikit-learn==1.4.2
streamlit==1.33.0
requests==2.31.0
watchdog==4.0.0

#pip show

